Start to reorganize.
